# docs/swagger.yaml
definitions:
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  models.StandardResponse:
    description: A standard API response format with data payload
    properties:
      data: {}
      error:
        example: Invalid input
        type: string
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  models.StandardResponseUser:
    description: A standard API response with user data
    properties:
      data:
        $ref: '#/definitions/models.SwaggerProfile'
      error:
        example: Invalid input
        type: string
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  models.SwaggerProfile:
    description: User profile information
    properties:
      bio:
        example: Software developer
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      profile_image:
        example: https://example.com/avatar.jpg
        type: string
      role:
        example: user
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  models.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        description: in seconds
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  models.TokenRevokeRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  models.UpdateProfileRequest:
    description: Request model for updating user profile
    properties:
      bio:
        example: Software developer
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT tokens
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Login to the application
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current user's tokens
      parameters:
      - description: Logout options
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Logout from the application
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Refresh an access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.StandardResponseUser'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "409":
          description: Email or username already exists
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Register a new user
      tags:
      - Auth
  /auth/revoke:
    post:
      consumes:
      - application/json
      description: Invalidate a refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/models.TokenRevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token revoked successfully
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Invalid input or token revocation failed
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Revoke a refresh token
      tags:
      - Auth
...
  /profile:
    get:
      description: Retrieve the current user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/models.SwaggerProfile'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the current user's profile information
      parameters:
      - description: Profile Data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.SwaggerProfile'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
...
swagger: "2.0"